ingestion/requirements.txt
finnhub-python==2.4.27
websocket-client==1.9.0
websockets==15.0.1
